Move the TARDISSpectrum class from model to montecarlo.spectrum.
